# Simplify `fill`