@seanmcl 2018-08-05
Hmm, I see.  The problem I was trying to solve with this one is that with hard line breaks, which do not exist in the paper, the existing `fill` method is rather difficult to specify.  As I was working on it I noticed that the existing invariants are not satisfied (#112).  Because we use three different occurrences of `Union` in the implementation, it's certainly one of the most complicated methods in the library.  I wondered if we could get away with making it a derived method (i.e. that only uses pre-defined combinators.)  Here I noticed that by using `lineOrSpace` instead of `line` we can achieve the same layout.  I don't know if that's true in general.  This change mimics `fillwords` from the paper, but I see that `fill` is different, so probably we just need more elaborate tests.  I'll try to come up with some that require `fill` being primitive.  In its existing form, `fill` is going to be a beast to change with hard breaks.