@seanmcl 2018-08-06
Dropping this.  Will make tests in another PR.